Changed getLink page
So I changed some of the responsiveness and made it so that the link
only appears if the link comes back from the url. pretty cool
Simple spell but quite unbreakable. I added a button so that the user
can copy their link to their clipboard instead of having to select the
link themselves. I also made it so that the link is hidden until it is
received from the api.